[MIG] base_tier_validation_correction: Migration to 19.0
