+++ base_tier_validation_correction/models/__init__.py
@@ -1,4 +1,5 @@
 # Copyright 2020 Ecosoft Co., Ltd. (http://ecosoft.co.th)
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
 from . import tier_correction
+from . import tier_correction_item
 from . import tier_validation
