+++ base_tier_validation_correction/models/tier_validation.py
@@ -2,6 +2,7 @@
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
 
 from odoo import models
+from odoo.fields import Domain
 from odoo.tools.safe_eval import safe_eval
 
 
@@ -11,7 +12,7 @@ class TierValidation(models.AbstractModel):
     def view_tier_correction(self):
         self.ensure_one()
         # Caller document
-        res_model = self.env["ir.model"].search([("model", "=", self._name)])
+        res_model = self.env["ir.model"].search(Domain("model", "=", self._name))
         # Prepare to redirect to Tier Correction
         action = self.env.ref("base_tier_validation_correction.tier_correction_action")
         result = action.sudo().read()[0]
@@ -29,11 +30,13 @@ class TierValidation(models.AbstractModel):
         )
         result["context"] = ctx
         items = self.env["tier.correction.item"].search_fetch(
-            [("res_model", "=", self._name), ("res_id", "=", self.id)],
+            Domain("res_model", "=", self._name) & Domain("res_id", "=", self.id),
             ["correction_id"],
         )
         if len(items) > 0:  # Already has corrections, show them.
-            result["domain"] = [("id", "in", items.mapped("correction_id").ids)]
+            result["domain"] = list(
+                Domain("id", "in", items.mapped("correction_id").ids)
+            )
         else:  # No correction, create one
             res = self.env.ref(
                 "base_tier_validation_correction.tier_correction_view_form", False
