+++ base_tier_validation_correction/security/correction_groups.xml
@@ -1,11 +1,15 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <odoo noupdate="0">
-    <record id="group_tier_correction" model="res.groups">
+    <record id="privilege_tier_correction" model="res.groups.privilege">
         <field name="name">Tier Review Correction</field>
         <field name="category_id" ref="base.module_category_hidden" />
+    </record>
+    <record id="group_tier_correction" model="res.groups">
+        <field name="name">Allowed</field>
+        <field name="privilege_id" ref="privilege_tier_correction" />
         <field
-            name="users"
-            eval="[(4, ref('base.user_root')), (4, ref('base.user_admin'))]"
+            name="user_ids"
+            eval="[Command.link(ref('base.user_root')), Command.link(ref('base.user_admin'))]"
         />
     </record>
 </odoo>
